requirements.txt
# Python 依赖
httpx>=0.27.0
playwright>=1.40.0
apscheduler>=3.10.4
python-dotenv>=1.0.0
psutil>=5.9.0  # 用于性能监控
pyyaml>=6.0    # 用于配置文件

# 数据库
sqlalchemy>=2.0.0

# Web 框架（第二阶段使用）
fastapi>=0.104.0
uvicorn>=0.24.0
jinja2>=3.1.2
python-multipart>=0.0.6
